Add files via upload
This version includes code for normality and homogeneity of variance tests before the t-test. Additionally, a new test method, permutation test, has been added. This method does not require averaging the pre-onset and post-onset data, nor does it require the sample sizes to be exactly equal. Instead, it randomly permutes the pre and post data to see if there is a statistical difference (here, a one-tailed test). It seems to have yielded a more reasonable result.